Move app file inside components folder
